Move code around, remove xfails
